# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  domain.ChangePasswordRequest:
    properties:
      current_password:
        type: string
      new_password:
        minLength: 6
        type: string
    required:
    - current_password
    - new_password
    type: object
...
  domain.Store:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        maxLength: 100
        minLength: 2
        type: string
      photo_url:
        type: string
      rating:
        type: number
      status:
        type: string
      updated_at:
        type: string
      user:
        allOf:
        - $ref: '#/definitions/domain.User'
        description: Relations
      user_id:
        type: integer
    required:
    - name
    type: object
...
  domain.UpdateProfileRequest:
    properties:
      date_of_birth:
        type: string
      name:
        maxLength: 100
        minLength: 2
        type: string
      phone:
        maxLength: 15
        minLength: 10
        type: string
    required:
    - name
    - phone
    type: object
  domain.UpdateStoreRequest:
    properties:
      description:
        type: string
      name:
        maxLength: 100
        minLength: 2
        type: string
    required:
    - name
    type: object
  domain.User:
    properties:
      about:
        type: string
      city_id:
        type: integer
      created_at:
        type: string
      date_of_birth:
        type: string
      email:
        type: string
      email_verified_at:
        type: string
      gender:
        type: string
      id:
        type: integer
      is_admin:
        type: boolean
      job:
        type: string
      last_login_at:
        type: string
      name:
        maxLength: 100
        minLength: 2
        type: string
      phone:
        maxLength: 15
        minLength: 10
        type: string
      photo_url:
        type: string
      province_id:
        type: integer
      status:
        type: string
      updated_at:
        type: string
    required:
    - email
    - name
    - phone
    type: object
  response.PaginatedResponse:
    properties:
      data: {}
      message:
        type: string
      meta:
        $ref: '#/definitions/response.PaginationMeta'
      status:
        type: string
    type: object
  response.PaginationMeta:
    properties:
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
      total_page:
        type: integer
    type: object
  response.Response:
    properties:
      data: {}
      message:
        type: string
      status:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: E-commerce API built with Go Fiber and Clean Architecture
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Go Commerce API
  version: "1.0"
paths:
...
  /stores:
    get:
      consumes:
      - application/json
      description: Get all stores with pagination and search (public endpoint)
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: limit
        type: integer
      - description: Search by store name
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Stores retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/response.PaginatedResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/domain.Store'
                  type: array
              type: object
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Get all stores
      tags:
      - Stores
  /stores/{id}:
    get:
      consumes:
      - application/json
      description: Get a single store by its ID (public endpoint)
      parameters:
      - description: Store ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Store retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/domain.Store'
              type: object
        "400":
          description: Invalid store ID
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Store not found
          schema:
            $ref: '#/definitions/response.Response'
      summary: Get store by ID
      tags:
      - Stores
  /stores/my:
    get:
      consumes:
      - application/json
      description: Get store information for the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: Store retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/domain.Store'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Store not found
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Get current user's store
      tags:
      - Stores
    put:
      consumes:
      - application/json
      description: Update store information for the authenticated user
      parameters:
      - description: Store update request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.UpdateStoreRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Store updated successfully
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/domain.Store'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Update current user's store
      tags:
      - Stores
...
  /users/change-password:
    put:
      consumes:
      - application/json
      description: Change current user password
      parameters:
      - description: Password change request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.ChangePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Password changed successfully
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Change user password
      tags:
      - Users
...
  /users/profile:
    get:
      consumes:
      - application/json
      description: Get current user profile information
      produces:
      - application/json
      responses:
        "200":
          description: Profile retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/domain.User'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Get user profile
      tags:
      - Users
    put:
      consumes:
      - application/json
      description: Update current user profile information
      parameters:
      - description: Profile update request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.UpdateProfileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Profile updated successfully
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/domain.User'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Update user profile
      tags:
      - Users
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"